+++ src/Relation/LoadBelongsTo.php
@@ -98,10 +98,10 @@ class LoadBelongsTo
         $query = $targetRepository->sqlQuery()
             ->whereIn($ownerKeyCol, $matchValues);
         if ($targetScope !== null) {
-            $targetScope($query, $children, $childRelation, $targetRepository, $sourceRepository);
+            $targetScope($query, $children);
         }
         if ($sourceFilter !== null) {
-            $sourceFilter($query, $children, $childRelation, $targetRepository, $sourceRepository);
+            $sourceFilter($query, $children);
         }
         $parents = $query->getResult();